refactor(LPX-600): rename `standalone` namespace to `solo`
`Standalone` was alpha-era language that never validated on a shipped
consumer. The name implies an architectural choice (standalone vs
embedded) but actually means "single tenant DNS + queue path." `Solo`
reads naturally as the sibling of `Multitenancy` and conveys what the
namespace actually represents.

Mechanical rename — zero behaviour changes, AWS resource names
unchanged. CLI subcommand `sync:standalone` → `sync:solo` (no
consumer invokes it directly; `sync` is the orchestrator).
