+++ src/Contracts/ExecutesSoloStep.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Codinglabs\Yolo\Contracts;
+
+interface ExecutesSoloStep extends Step {}
