--- src/Contracts/ExecutesStandaloneStep.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Codinglabs\Yolo\Contracts;
-
-interface ExecutesStandaloneStep extends Step {}
