+++ src/Steps/Solo/SyncHostedZoneStep.php
@@ -1,33 +1,29 @@
 <?php
 
-namespace Codinglabs\Yolo\Steps\Standalone;
+namespace Codinglabs\Yolo\Steps\Solo;
 
 use Codinglabs\Yolo\Aws;
 use Illuminate\Support\Arr;
-use Codinglabs\Yolo\Helpers;
+use Illuminate\Support\Str;
+use Codinglabs\Yolo\Manifest;
 use Codinglabs\Yolo\AwsResources;
-use Codinglabs\Yolo\Contracts\Step;
 use Codinglabs\Yolo\Enums\StepResult;
-use Codinglabs\Yolo\Contracts\ExecutesStandaloneStep;
+use Codinglabs\Yolo\Contracts\ExecutesDomainStep;
 use Codinglabs\Yolo\Exceptions\ResourceDoesNotExistException;
 
-class SyncQueueStep implements ExecutesStandaloneStep, Step
+class SyncHostedZoneStep implements ExecutesDomainStep
 {
     public function __invoke(array $options): StepResult
     {
-        $name = Helpers::keyedResourceName();
-
         try {
-            AwsResources::queue($name);
+            AwsResources::hostedZone(Manifest::apex());
 
             return StepResult::SYNCED;
         } catch (ResourceDoesNotExistException) {
             if (! Arr::get($options, 'dry-run')) {
-                Aws::sqs()->createQueue([
-                    'QueueName' => $name,
-                    'Attributes' => [
-                        'MessageRetentionPeriod' => '1209600', // 14 days
-                    ],
+                Aws::route53()->createHostedZone([
+                    'CallerReference' => Str::uuid(),
+                    'Name' => Manifest::apex(),
                 ]);
 
                 return StepResult::CREATED;
