+++ src/Steps/Solo/SyncQueueStep.php
@@ -1,29 +1,33 @@
 <?php
 
-namespace Codinglabs\Yolo\Steps\Standalone;
+namespace Codinglabs\Yolo\Steps\Solo;
 
 use Codinglabs\Yolo\Aws;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Str;
-use Codinglabs\Yolo\Manifest;
+use Codinglabs\Yolo\Helpers;
 use Codinglabs\Yolo\AwsResources;
+use Codinglabs\Yolo\Contracts\Step;
 use Codinglabs\Yolo\Enums\StepResult;
-use Codinglabs\Yolo\Contracts\ExecutesDomainStep;
+use Codinglabs\Yolo\Contracts\ExecutesSoloStep;
 use Codinglabs\Yolo\Exceptions\ResourceDoesNotExistException;
 
-class SyncHostedZoneStep implements ExecutesDomainStep
+class SyncQueueStep implements ExecutesSoloStep, Step
 {
     public function __invoke(array $options): StepResult
     {
+        $name = Helpers::keyedResourceName();
+
         try {
-            AwsResources::hostedZone(Manifest::apex());
+            AwsResources::queue($name);
 
             return StepResult::SYNCED;
         } catch (ResourceDoesNotExistException) {
             if (! Arr::get($options, 'dry-run')) {
-                Aws::route53()->createHostedZone([
-                    'CallerReference' => Str::uuid(),
-                    'Name' => Manifest::apex(),
+                Aws::sqs()->createQueue([
+                    'QueueName' => $name,
+                    'Attributes' => [
+                        'MessageRetentionPeriod' => '1209600', // 14 days
+                    ],
                 ]);
 
                 return StepResult::CREATED;
